feat: move current tickets to separate components, workers can update status to in-progress or finished
